// src/api/model/alarmLog.ts
/**
 * LinkIOT Dashboard
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface AlarmLog { 
    id: number;
    deviceId: string;
    data: string;
    createTime: string;
    level: number;
    handle: string;
    status: Boolean;
}